completely refactored project structure
