+++ src/render/mod.rs
@@ -1,0 +1,8 @@
+mod camera;
+mod color;
+mod menu;
+pub(crate) mod message;
+pub(crate) mod pipeline;
+mod polyblade;
+mod state;
+pub use polyblade::Polyblade;
